app/requirements.txt
aiosignal==1.3.1
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorclass==2.2.2
contourpy==1.0.7
cramjam==2.6.2
cryptography==40.0.1
cycler==0.11.0
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
easygui==0.98.3
fastparquet==2023.4.0
fasttext==0.9.2
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
greenlet==2.0.2
grpcio==1.53.0
idna==3.4
img2pdf==0.4.4
importlib-resources==5.12.0
iso639==0.1.4
joblib==1.2.0
jsonlines==3.1.0
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
msgpack==1.0.5
msoffcrypto-tool==5.0.1
numpy==1.21.5
olefile==0.46
oletools==0.60.1
opencv-python==4.5.5.64
packaging==23.0
pandas==1.5.1
pcodedmp==1.2.6
pdf2image==1.16.0
pdfminer.six==20221105
pdfplumber==0.8.1
pikepdf==7.1.2
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
protobuf==4.22.1
psutil==5.9.4
pybind11==2.10.4
pycparser==2.21
pyparsing==2.4.7
pyrsistent==0.19.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
PyYAML==6.0
regex==2021.11.10
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
tabulate==0.9.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
py4j==0.10.9.5
warcio==1.7.4
pyarrow==12.0.0
gitpython==3.1.32
ray
